requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
datasets>=2.14.0
langchain>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0
astor>=0.8.1
pylint>=3.0.0
pyyaml>=6.0.1
wandb  # optional logging
trl>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
sentencepiece
protobuf

